boot: install boot config to /usr/share/boot
Rather than installing the boot configuration files directly to the
/boot partition, install them to /usr/share/boot and then copy them
to /boot in the post-install script.

This is necessary because dpkg creates a hard link to back-up an
existing file when upgrading, and the /boot partition is formatted
as FAT, which doesn't support hard links.

Wrote a README with basic instructions for building.

added GPL notice to package.sh